refactor: post 도메인과 source 도메인의 fixture 소유권 정리
