--- src/test/java/com/techfork/domain/source/fixture/SourcePostFixture.java
@@ -1,45 +1,0 @@
-package com.techfork.domain.source.fixture;
-
-import com.techfork.post.domain.Post;
-import com.techfork.domain.source.dto.RssFeedItem;
-import com.techfork.domain.source.entity.TechBlog;
-import org.springframework.test.util.ReflectionTestUtils;
-
-import java.time.LocalDateTime;
-
-public final class SourcePostFixture {
-
-    private static final LocalDateTime DEFAULT_PUBLISHED_AT = LocalDateTime.of(2026, 4, 13, 7, 0, 0);
-    private static final LocalDateTime DEFAULT_CRAWLED_AT = LocalDateTime.of(2026, 4, 13, 8, 0, 0);
-
-    private SourcePostFixture() {
-    }
-
-    public static Post createPost(Long id, String title, String fullContent, String plainContent,
-                                  String company, String summary, String shortSummary) {
-        TechBlog techBlog = TechBlog.create(
-                company,
-                "https://%s.example.com".formatted(company.toLowerCase()),
-                "https://%s.example.com/rss".formatted(company.toLowerCase()),
-                "https://cdn.example.com/%s.png".formatted(company.toLowerCase())
-        );
-        ReflectionTestUtils.setField(techBlog, "id", id);
-
-        Post post = Post.create(RssFeedItem.builder()
-                .title(title)
-                .url("https://posts.example.com/%s".formatted(id))
-                .logoUrl("https://cdn.example.com/logo-%s.png".formatted(id))
-                .thumbnailUrl("https://cdn.example.com/thumb-%s.png".formatted(id))
-                .content(fullContent)
-                .plainContent(plainContent)
-                .publishedAt(DEFAULT_PUBLISHED_AT)
-                .company(company)
-                .techBlogId(id)
-                .build(), techBlog);
-        ReflectionTestUtils.setField(post, "id", id);
-        ReflectionTestUtils.setField(post, "crawledAt", DEFAULT_CRAWLED_AT);
-        ReflectionTestUtils.setField(post, "summary", summary);
-        ReflectionTestUtils.setField(post, "shortSummary", shortSummary);
-        return post;
-    }
-}
