+++ src/test/java/com/techfork/post/domain/PostTest.java
@@ -2,6 +2,8 @@ package com.techfork.post.domain;
 
 import com.techfork.domain.source.dto.RssFeedItem;
 import com.techfork.domain.source.entity.TechBlog;
+import com.techfork.domain.source.fixture.RssFeedItemFixture;
+import com.techfork.domain.source.fixture.TechBlogFixture;
 import com.techfork.post.fixture.PostFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -21,9 +23,9 @@ class PostTest {
         @Test
         @DisplayName("RssFeedItem과 TechBlog로 게시글 aggregate를 생성한다")
         void createsPostAggregateFromRssFeedItem() {
-            TechBlog techBlog = PostFixture.createTechBlog(1L, "TechFork");
+            TechBlog techBlog = TechBlogFixture.createTechBlog(1L, "TechFork");
             LocalDateTime publishedAt = LocalDateTime.of(2026, 5, 7, 10, 30);
-            RssFeedItem rssFeedItem = PostFixture.createRssFeedItem(
+            RssFeedItem rssFeedItem = RssFeedItemFixture.createRssFeedItem(
                     1L,
                     "Post aggregate 설계",
                     "https://posts.example.com/post-aggregate",
